refactor: split validators tabs
